createSchedule UI
My part of the UI for creating new schedules for an autmated embedded
system. Could probably compile and run via npm commands, but lacks
fuctionality without the rest of the project.

The actual implemented project can be found here:
http://autowater2.knewrealitys.com:8081/login

However, the project is being used personally by the other members and
authentification is required.